migrate barbar to lazy.nvim
